Move login helpers to a separate file
